RR-231: Prepare to move details into this space Remove duplication Delete unused code in app space
