+++ src/apps/events/client/Details/tasks.js
@@ -1,7 +1,7 @@
 import { apiProxyAxios } from '../../../../client/components/Task/utils'
-import { transformResponseToEventCollectionDetails } from '../transformers'
+import { transformResponseToEventDetails } from '../../../../client/modules/Events/transformers'
 
 export const getEventDetails = (eventId) =>
   apiProxyAxios
     .get(`/api-proxy/v3/event/${eventId}`)
-    .then(({ data }) => transformResponseToEventCollectionDetails(data))
+    .then(({ data }) => transformResponseToEventDetails(data))
