--- src/apps/events/client/constants.js
@@ -1,32 +1,0 @@
-export const LABELS = {
-  eventName: 'Event name',
-  organiser: 'Organiser',
-  startDateAfter: 'From',
-  startDateBefore: 'To',
-  country: 'Country',
-  ukRegion: 'UK region',
-  eventType: 'Type of event',
-}
-
-export const SORT_OPTIONS = [
-  {
-    name: 'Event name A-Z',
-    value: 'name:asc',
-  },
-  {
-    name: 'Recently updated',
-    value: 'modified_on:desc',
-  },
-  {
-    name: 'Least recently updated',
-    value: 'modified_on:asc',
-  },
-  {
-    name: 'Earliest start date',
-    value: 'start_date:asc',
-  },
-  {
-    name: 'Latest start date',
-    value: 'start_date:desc',
-  },
-]
